requirements.txt
fastapi==0.115.12
pytest==8.4.0
python-telegram-bot==22.1
uvicorn==0.34.3

python-dotenv~=1.1.0